Sync autoTick and the current xTickInterval and yTickInterval with the OTDataGraph.
git-svn-id: https://svn.concord.org/svn/projects/trunk/common/java/data/datagraph@23967 6e01202a-0783-4428-890a-84243c50cc2b

+++ src/main/java/org/concord/datagraph/state/OTDataGraph.java
@@ -91,13 +91,13 @@ public interface OTDataGraph extends OTObjectInterface
 	/**
 	 * This set the tick interval. This is read ONLY if autoTick is false.
 	 */
-	public int getXTickInterval();
-	public void setXTickInterval(int xTickInterval);
-	public static int DEFAULT_xTickInterval = 2;
+	public double getXTickInterval();
+	public void setXTickInterval(double xTickInterval);
+	public static double DEFAULT_xTickInterval = 2;
 	
-	public int getYTickInterval();
-	public void setYTickInterval(int yTickInterval);
-	public static int DEFAULT_yTickInterval = 2;
+	public double getYTickInterval();
+	public void setYTickInterval(double yTickInterval);
+	public static double DEFAULT_yTickInterval = 2;
     
 	/** This determines whether the list of graphables in shown to the left of the graph */
 	public boolean getShowGraphableList();
